sui_data_store

Trait EpochStore

pub trait EpochStore {
    // Required methods
    fn epoch_info(&self, epoch: u64) -> Result<Option<EpochData>, Error>;
    fn protocol_config(
        &self,
        epoch: u64,
    ) -> Result<Option<ProtocolConfig>, Error>;
}

An EpochStore retrieves the epoch data and protocol configuration for a given epoch. Epoch data is collected by an indexer and it is not stored anywhere otherwise. This is a very small amount of information and could conceivably be saved locally and never hit a server.

Required Methods

fn epoch_info(&self, epoch: u64) -> Result<Option<EpochData>, Error>

Return the EpochData for a given epoch.

fn protocol_config(&self, epoch: u64) -> Result<Option<ProtocolConfig>, Error>

Return the ProtocolConfig for a given epoch.