sui_data_store

Trait TransactionStore

pub trait TransactionStore {
    // Required method
    fn transaction_data_and_effects(
        &self,
        tx_digest: &str,
    ) -> Result<Option<TransactionInfo>, Error>;
}

A TransactionStore has to be able to retrieve transaction data for a given digest. The data provided to sui_execution::executor::Executor::execute_transaction_to_effects must be available. Some of that data is not provided by the user. It is naturally available at runtime on a live system and later saved in effects and in the context of a checkpoint.

Required Methods

fn transaction_data_and_effects( &self, tx_digest: &str, ) -> Result<Option<TransactionInfo>, Error>

Given a transaction digest, return transaction info including data, effects, and the checkpoint that transaction was executed in. Returns None if the transaction is not found.